+++ apps/doc/src/app/version-manager/versions.constants.ts
@@ -6,7 +6,7 @@ export interface PrizmVersionMeta {
 
 export const PRIZM_VERSIONS_META: readonly PrizmVersionMeta[] = [
   {
-    label: '2.1.4',
+    label: '2.1.5',
     baseHref: 'http://prizm.site/',
   },
   {