+++ libs/components/src/lib/modules/overlay/utils.ts
@@ -24,7 +24,7 @@ export function getContent(
 }
 
 /* html dom utils */
-export function cssClass(method: 'add' | 'remove', cls: string[], target: string = 'body'): void {
+export function cssClass(method: 'add' | 'remove', cls: string[], target = 'body'): void {
   document.querySelector(target).classList[method](...cls);
 }