+++ apps/doc/src/app/about-prizm/changelog/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.4.3-next.1, 2.1.5-next.1, 3.0.0-next.1](https://github.com/zyfra/Prizm) (17-08-2023)
+
+### Features
+
+- feat(nxmv): published new library to support multiple dependencies (ng) #595
+- feat(components/sidebar): overlay with parent #576
+
 ## [1.4.2](https://github.com/zyfra/Prizm) (14-08-2023)
 
 ### Features